Add some Fedora Mobile magic.
If given `?fedora_mobile=true` on a profile page, we render a link that
the Fedora Mobile app will eventually know how to deal with. I think
this is fine to add now (before Fedora Mobile actually supports it),
because without adding `?fedora_mobile=true`, nobody will ever see it.