Queue and flush messages when lost client.
This will queue messages indefinitely when the irc backend has no
clients to work with.  This is typically the case when the bot has
suffered a netsplit, or when it is first starting up but hasn't logged
in yet.

Later, if the connection is regained, the next message will prompt the
bot to flush queued messages.

Fixes #12.