This thing doesn't actually have access to the config.
This was causing a bug where the bot would query NickServ, NickServ would
respond.  The irc bot would traceback, and try again.  We spammed NickServ over
and over again, unable to "hear" its reply.

Not sure when this got introduced, but this is a backport from a hotfix -- it
is definitely hammered out in production at this point.