Reduce spam.
This adds a notify boolean to one of the Preference modifiers that will let us
turn off fedmsg messages for all but one of the initial operations that goes
into creating a new account.  This should reduce spam on the message bus.

This is especially important because these messages cause the fmn backend to
*invalidate its own cache*, thus slowing it down dramatically.  There is no
need for it to invalidate its own cache *20 times* for a new user.  It should
only need to do it once.