Fix the "examples" feature.
There's this feature where if you go to look at a particular **rule** in fmn,
then it will asynchronously query datagrepper for messages that *would have
matched* that rule, so you can get a feel for what it would be like to have it
in place.

This used to work, but it broke in the [last
refactor](http://threebean.org/blog/optimizing-fmn/).  This fixes it again.