Don't return prematurely.
It made it so that as soon as one rule matched for one user, the whole tamale
would evaluate to true for every subsequent user.  This incorrectly made it so
that users would suddenly start getting all kinds of messages not relevant to
them; messages would be directed at 40+ users simultaneously, one after
another, after another.