Less formal short-descriptions for rules.
It used to be that all the short-descriptions of the rules in the web ui
appeared very formal.  They were prefixed with some identifier like
"Fedora Hosted: ....".

This changeset tries to make them a little less formal and a little more
human-readable.  The long descriptions are still present so users can
click on them to get more info.

The intent is to have the UI read more fluidly like this:

- *Include* A particular user's packages (ralph's)
- *Include* Build state changes
- *Ignore*  Builds starting