Indicate to users that they should check their email.
@ryanlerch hit this and reported it in IRC.

When people first enter their email address, it is in the 'pending'
state.

Then, moments later, a polling producer in a backend fedmsg process
notices that they said they want to use this email address and it tries
to email them with a confirmation link.  When it *sends* that email it
moves their address from the 'pending' state to the 'valid' state (but
not yet 'confirmed').

Finally, when they click the link they get, it goes to the 'confirmed'
state.

Ryan hit a race condition where his page reloaded slower than it took
that backend process to email him, so he never got the warning that he
should be going to check his email.

![](http://i.imgur.com/CWazPQL.png)

This change keeps that notice there on the page for both the pending and
valid states.. to make things more clear.