Ditch old re2 warning hook.
Google has a regular expression lib in C called 're2'.

There are multiple python bindings for it out there.  We used to use the one by
axiak, called pyre2 on pypi.  At PyCon, I learned that his bindings are
unmaintained, incomplete, and incorrect.  They just so happened to work for the
case we were using it for in fmn.rules, so we didn't notice.

Facebook engineers maintain a more up-to-date set of python bindings called
fb-re2 on pypi.  They approached us at PyCon and we swapped out the old pyre2
for fb-re2 in the ``python-re2`` module in EPEL and Fedora.

Anyways.. we're updated to that new fb-re2 source and it doesn't produce this
little warning function that we were calling, so just drop it.