Queue or not the message that raised an exception
When running into smtplib.SMTPSenderRefused, most often caused by the
message being to big, we just nack the message and don't requeue it as
this message will always be too big and therefore the user will never
see it.

When running into RuntimeError, which we raise when there are no clients
connected to IRC, requeue the message so that when the clients re-connect
the message can be sent.