Convert the SSE server to a Twisted pika connection
Prior to this commit, a blocking connection to RabbitMQ was used. This
is problematic in Twisted because it blocks the reactor rather than take
advantage of Twisted's reactor design.

This should behave almost exactly like the previous connection, except
that now messages are pushed by RabbitMQ to the SSE server and in turn
immediately pushed out to clients. It also works with a single TCP
connection rather than using one connection per queue.

Fixes #48 and #50

Signed-off-by: Jeremy Cline <jeremy@jcline.org>