Add a call to stop consuming from the AMQP queue
Closing the ``queue`` the Twisted Pika connection provides does not
automatically send a message to the AMQP server to stop consuming from a
queue. This commit adds that call.

Note that when the client closes the connection to the Twisted server,
the error handling doesn't kick in until the next message is processed.
If there are no messages in the queue, the consumer will not cancel its
subscription until a message is pushed to the consumer which triggers
the error handling function.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>