Add a Twisted application file and configure logging
Twisted uses its own (non-blocking) logging system. It can be configured
in simple ways using the ``twistd`` tool used to launch the application
which should suffice for now. However, there are third-party libraries
that use the standard Python logging module, so this is configured as
part of the application's setup.

fixes #58

Signed-off-by: Jeremy Cline <jeremy@jcline.org>