Move fmn.sse out of the fmn namespace
The fmn.sse setup.py was not declaring itself as part of the fmn
namespace, nor was the __init__.py setting it up. On reflection, I don't
see much value in this being part of the fmn namespace in any case since
it is unlikely to ever be used by anything else in the fmn package.
Furthermore, package namespaces are only natively supported in Python in
3.3+ and setuptools handles the backwards compatibility in earlier
versions. Since it was so easy to move it out of the fmn namespace I've
gone ahead and done this here.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>