Start creating a scoped session object in the models module
Since we load the configuration at module import time anyway, there's no
reason we can't create the database engine at import time as well. The
engine only connects to the database when connect is explicitly called
or something that requires a connection is called[0]. Furthermore, we
can also create the scoped session object.

This adds the ability to create the database to the fmn-createdb script
and stops that from calling ``fmn.lib.models.init``. That function has
been deprecated.

[0] http://docs.sqlalchemy.org/en/latest/core/engines.html#sqlalchemy.create_engine

Signed-off-by: Jeremy Cline <jeremy@jcline.org>