Stop trying to shuffle preferences in the worker consumer
When combining the various load_preferences functions in #175, a call of
``random.shuffle`` was overlooked in the worker (where preferences
stopped being a list and started being a dictionary like in the
backend). It's obviously not possible to shuffle a dictionary, and the
purpose of the call, which appears to exist to avoid resource contention
from different threads, no longer applies.

fixes #181

Signed-off-by: Jeremy Cline <jeremy@jcline.org>