Cache a mapping of rule strings to rule objects
The preference loading resulted in loading rule objects via
``__import__`` for each rule in the database. For 170 users with default
preferences, this is approximately 30,000 rows and therefore 30,000
imports. Using a dictionary cache decreased the load time by nearly an
order of magnitude (in decimal) from 56 seconds to ~6 seconds. With
larger rulesets this should increase by even more than an order of
magnitude as there will be a larger percentage of cache hits.

This is a memory-backed cache that never prunes its entries. However,
there are currently only about 300 rules in FMN and that is not a
particularly large dictionary.

fixes #189

Signed-off-by: Jeremy Cline <jeremy@jcline.org>