Cache the rule-loading code
``fmn.lib.load_rules`` is called all over the place without arguments,
and each time it takes about 2.5 seconds to load (on a fast disk) before
it's calling ``__import__`` on tons of files.

This is a short-term fix that simply caches the results in a dictionary.
This makes creating 100 users drop from 221 seconds to 3.3 seconds.

fixes #191

Signed-off-by: Jeremy Cline <jeremy@jcline.org>