Ensure all the dogpile.cache regions use a copy of the config
Older versions of dogpile.cache mutate the configuration provided to the
``configure`` call. This is problematic since we currently configure
several regions with the same dictionary. This makes sure each provide a
copy of the dictionary.

Longer term, all the configuration should happen lazily, but that's a
big refactor.

Finally, as much as I would like to write tests for this, too much stuff
is happening in FMN at import-time so unfortunately, it's not really
possible to test this.

fixes #194

Signed-off-by: Jeremy Cline <jeremy@jcline.org>