Switch to Celery broadcasts for cache invalidation
Workers need to refresh their caches when we get messages about user
preference changes. This sets up a broadcast for workers.

There are a few oddities with this patch. Firstly, because of the way
the cache is initialized (inside the Task instance) we have to either dig
around in there to update the cache or dispatch the task itself and do a
special check to see if it's a cache refresh task. In this patch I've
opted for the second approach, although I'm not thrilled about that. In
the future we should have a better caching layer that makes this less
painful. Secondly, workers are being started in --pool=solo mode (one
process per worker). This is because the broadcast gets sent to each
worker, but not to each process in the worker's process pool. Running
in solo mode maps workers to processes 1-to-1, although it is generally
less efficient to run that way.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>