Convert the FMN backend to a Twisted application
Prior to this commit, the backend portion of FMN was a Python module
which ran the service as a side effect of importing.

This changes several pieces of the backend:

* The backend service is now a Twisted application (run with twistd)

* The digest service is now a celery task dispatched with celery beat
  rather than a copy/paste of the backend service with some changes.

* The fmn workers are responsible for formatting the messages for
  delivery, not the backend. Thus, the backend API has changed.

* There is a now a formatters module where all the formatting code
  lives. This was pulled out of the various backends.

* Use the Twisted email API rather than the stdlib one.

* I wrote some docs.

This commit is enormous and does several things, but all those things
were rather tangled together before and it would be a lot of work to
make this into a series of atomic commits. Given that FMN doesn't really
work at the moment, I've opted to just do one really big commit.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>