Move confirmation generation to a celery beat task
Move the database querying and message formatting to a celery task
dispatched by celery beat rather than a looping call in Twisted.

The advantage is two-fold. It makes confirmations not be a special case
(except for IRC at the moment, but that will take a substantial
refactor) and it removes the regular blocking IO from the Twisted
thread, improving throughput on the delivery service.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>