Just nack and requeue the message on failure in delivery
Previously, the delivery service formatted the messages and regularly
hit messages it couldn't render due to issues with fedmsg.meta. Now,
however, all the message formatting (except for confirmations in the IRC
backend which don't use fedmsg.meta) occurs on the workers when the
message recipients are calculated. This means that the delivery service
should only fail when a) there's a bug in the code where messages can't
be delivered at all or b) there's a network partition so we're not
connected to the mail server/irc/etc.

Because of this, we can stop dequeuing and manually republishing the
message to make it go to the back of the queue. In theory, any exception
should occur regardless of the message. It would be even better to
separate delivery backend messages into separate queues so that if IRC
is unreachable, it doesn't congest email.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>