Improve Pagure querying
We want all the packages for a user so we should query the largest amount
possible to avoid tons of tiny queries. Additionally, we should use a
session since we're going to make a lot of connections to the same host.
Finally, the API returns 404 when the query returns an empty collection
so just ignore that instead of logging it.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>