Fix the last failing test in tox.
It looks like dogpile.cache changed its behavior which resulted in the
tests passing locally via py.test (running dogpile.cache 0.9.0) and
failing in tox (running dogpile.cache 1.1.2).
Us manually populating the cache at the start of the test no longer
worked. So rather than trying to manually populate the cache, we're
calling the function twice, the first run should populate the cache
the second should retrieve from it.
So we can check afterward that the function used to populate the cache
was only called once.
This works both in py.test locally as well as in tox.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>