Improve FastAPI dependency for DB sessions
The implementation is good but its name was awkward. As it's only
intended for the API service, there's no need to explicitly state it is
for use in request handlers and yields async session. On the other hand,
having a doc string should make it clearer what the function is about.

Also, add tests.

Signed-off-by: Nils Philippsen <nils@redhat.com>