Fine tune the multiselect settings
This really changes 3 things:

1. Previously the text when there was no options (i.e. the user hasnt
   typed anything or the async search hasnt returned anything) was an
unhelpful "the list is empty" this passes the "noOptionsText" param
through to the multiselect, so we can set this tring to something a
little more helpful, which we also do for tyhe user search and the
artifact search.

2. We hadd some more helpful placeholders for the user search and
   artifact search.

3. finally we add the "clearOnSearch" to the multiselect. This clears
   all the options when we do a new search. This is primarily to stop
the multiselect from showing the list of all options previously loaded
when removing text.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>