Isolate unit tests from network
We use the `responses` package to mock out HTTP traffic to external
systems which use the `requests` package. However, FASJSON calls were
only configured to be mocked out for some tests, others used passthrus
which let them communicate with the Fedora FASJSON service and
json-schema.org. This fails every now and then when there's some hickup
on the network or one of the involved systems.

This consolidates mocking out requests to FASJSON and related services
in the `mocked_fasjson` fixture, and extends `mocked_fasjson_client` so
it can be used anywherre and makes both available globally.

Fixes: #635

Signed-off-by: Nils Philippsen <nils@redhat.com>