Fix getting projects by group
Previously, we were using the following distgit/pagure endpoint to
return a list of projects that a group has commit access to:

/api/0/projects?username=@mygroupname

while this does work to return a list of projects that a user has commit
access to (without the @ prefix) i can't find any mention of the @
prefix to denote a group in the pagure docs, and testing thge API with
this endpoint and the @ prefix returned nothing.

Therefore, this PR re-implements this endpoint to use:

/api/0/group/mygroupname?projects=true

to return a list of projects the group has commit access to.

Additionally, note that this changes it in the sync rules code, as well
as the async API code.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>