Don’t block API when sending bus messages
Bus messages are sent in a FastAPI background task, which are executed after an
API call has been serviced. However, this still blocks the (implicit) DB
commit from the `gen_db_session()` dependency as `fedora_messaging` is
synchronous code.

With this change, the synchronous code is run in the FastAPI thread pool
and not being awaited.

In the course, move logging the final exception when backoff attempts
are exceeded to a handler function instead of dealing with it on our
own.

Fixes: #678

Signed-off-by: Nils Philippsen <nils@redhat.com>