Fix flakiness around testing messages sent to bus
Previously, messages were effectively sent synchronously, i.e. the API
call would block until that is done, which made using `mock_sends()` of
`fedora_messaging` possible. PR #654 makes publishing messages
non-blocking in the API and as a result they are no longer guaranteed to
be sent before a request is serviced, so we have mock out the call in
the API handler and check if it is called instead.

Additionally, there’s no need to use FastAPI background tasks anymore,
so get rid of that level of indirection.

Related: #621, #654
Fixes: #684

Signed-off-by: Nils Philippsen <nils@redhat.com>