Compatibility with Fedora Messaging >= 3.3.0
The yet unreleased version of Fedora Messaging uses the asyncio reactor.
As a result, we should expect that the event loop is already running.

When F-M 3.3.0 is out, we can switch the Consumer's `__call__()` method
to an async method, and we'll be able to drop this compatibility layer.
In the meantime, we should expect to still be run in a threadpool.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>