Generate the typescript types from the OpenAPI spec
Use [openapi-typescript](https://github.com/drwpow/openapi-typescript)
to generate the Typescript types corresponding to the API schema as
generated by FastAPI. And use them in the frontend.

This ensures a pretty good match between the API and the frontend. We
must just remember to update the generated file when the API types
change. It should be done as part of the tests in CI, but it won't
commit the changes if there are any. It would be a good use case for a
pre-commit hook, but we're not using it (yet).

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>