Reuse proxy objects
This avoids creating essentially the same objects over and over again.
As a result, this speeds up __str__() for these objects because its
value is already cached, in order to speed up cashews cache key
computation.

On the consumer side, this is done implicitly already: Fedora Messaging
instantiates the Consumer once, which aggregates the Requester,
which in turn aggregates FASJSON and DistGit/Pagure proxy objects.

Fixes: #778

Signed-off-by: Nils Philippsen <nils@redhat.com>