Bring back flake8 configuration
This was removed when we switched to using Ruff instead of a collection
of individual linter and static analysis tools.

The reason to bring it back is that flake8 doesn’t complain about overly
long lines, if it’s configured to be run e.g. by the editor or a global
pre-commit hook.

Signed-off-by: Nils Philippsen <nils@redhat.com>