Don't use the early caching mechanism, because it leaks DB connections
Instead of lasting for the duration of the message handling, the DB
connection is used in the fire-and-forget cache rebuild system, which
can last quite a while.

If we want to rebuild the cache early, we should instead run a cron job
that checks cache freshness and rebuilds if necessary.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>