Filter unexpected project namespaces from Pagure
The API endpoints proxying Pagure validate their results but querying
Pagure could yield project namespaces not known to FMN (by mistake or
new kinds of artifacts introduced) which makes the handler croak.

This was found to be necessary because of a bug in tinystage, where RPM
package projects in test data used the `rpm` namespace instead of
`rpms`. See fedora-infra/tiny-stage#44.

Fixes: #831

Signed-off-by: Nils Philippsen <nils@redhat.com>