Only return one name variation per application
Version 0.9 of copr-messaging has some messages declare their app_name
as `Copr`, some not declaring it. This led to both `Copr` and `copr` be
returned from the /applications endpoint. This changes it so the variant
with most leading capital letters wins over others.

Additionally, cache the result, it’s likely to be queried often and
shouldn’t change over the runtime of the API backend process.

Related: #854

Signed-off-by: Nils Philippsen <nils@redhat.com>