Rebuild the cache instead of deleting it
When receiving a message that suggests we should invalidate the cache,
we will now rebuild it in the background instead of deleting it.
Rebuilding takes a lot of time, and if we delete the cache it will block
the processing of incoming messages.

We prefer working with a stale cache for some amount of time rather than
blocking message processing for that same amount of time.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>